backend/requirements.txt
# Backend
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
pymongo>=4.6
transformers>=4.38
torch>=2.1
faiss-cpu>=1.7
numpy>=1.24
nltk>=3.8
scikit-learn>=1.3
